clean up WDUI
