+++ beta-src/src/components/ui/WDPhaseUI.tsx
@@ -1,7 +1,5 @@
 import * as React from "react";
-import { Box } from "@mui/material";
 import {
-  gameOrdersMeta,
   gameOverview,
   gameStatus,
   gameViewedPhase,
@@ -13,6 +11,7 @@ import {
 } from "../../utils/state/getPhaseSeasonYear";
 import WDCountdownPill from "./WDCountdownPill";
 import WDPillScroller from "./WDPillScroller";
+import { abbrMap } from "../../enums/Country";
 
 const getCurPhaseMinutes = function (phaseMinutes, phaseMinutesRB, phase) {
   if (phaseMinutesRB !== -1 && (phase === "Retreats" || phase === "Builds")) {
@@ -31,9 +30,10 @@ const WDPhaseUI: React.FC = function (): React.ReactElement {
     year,
     processStatus,
     pauseTimeRemaining,
+    user,
   } = useAppSelector(gameOverview);
   const gameStatusData = useAppSelector(gameStatus);
-  const { viewedPhaseIdx, latestPhaseViewed } = useAppSelector(gameViewedPhase);
+  const { viewedPhaseIdx } = useAppSelector(gameViewedPhase);
 
   const phaseSeconds =
     getCurPhaseMinutes(phaseMinutes, phaseMinutesRB, phase) * 60;
@@ -48,16 +48,6 @@ const WDPhaseUI: React.FC = function (): React.ReactElement {
     season: viewedSeason,
     year: viewedYear,
   } = getHistoricalPhaseSeasonYear(gameStatusData, viewedPhaseIdx);
-  const ordersMeta = useAppSelector(gameOrdersMeta);
-  const ordersMetaValues = Object.values(ordersMeta);
-  const ordersLength = ordersMetaValues.length;
-  const ordersSaved = ordersMetaValues.reduce(
-    (acc, meta) => acc + +meta.saved,
-    0,
-  );
-  const animateForwardGlow =
-    latestPhaseViewed < gameStatusData.phases.length - 1 ||
-    ordersSaved !== ordersLength;
 
   // On the very last phase of a finished game, webdip API might give an
   // entirely erroneous year/season/phase. So instead, trust the one in the
@@ -72,23 +62,18 @@ const WDPhaseUI: React.FC = function (): React.ReactElement {
   const seconds = Math.floor(new Date().getTime() / 1000);
 
   return (
-    <Box
-      sx={{
-        display: "flex",
-        flexWrap: "wrap",
-        webkitFlexWrap: "wrap",
-        alignItems: "center",
-        height: 40,
-        marginTop: "3px",
-        marginRight: "55px", // for right side icons
-        pointerEvents: "none", // this box is invisible and used for layout alone, it shouldn't mask out clicks behind it
-      }}
-    >
+    <div className="flex flex-col space-y-2">
+      {user && (
+        <div
+          className={`py-1 px-2 rounded-md w-fit font-medium select-none text-white bg-${user.member.country.toLocaleLowerCase()}-main`}
+        >
+          {/* title={`Currently playing as ${user.member.country}`} */}
+          {abbrMap[user.member.country]}
+          <div className="hidden bg-france-main bg-austria-main bg-england-main bg-germany-main bg-russia-main bg-italy-main bg-turkey-main" />
+        </div>
+      )}
       <WDPillScroller
-        backwardDisabled={viewedPhaseIdx === 0}
-        forwardDisabled={viewedPhaseIdx === gameStatusData.phases.length - 1}
-        animateForwardGlow={animateForwardGlow}
-        viewedPhase={viewedPhase}
+        country={user?.member.country || ""}
         viewedSeason={viewedSeason}
         viewedYear={viewedYear}
       />
@@ -105,8 +90,10 @@ const WDPhaseUI: React.FC = function (): React.ReactElement {
           isPaused={processStatus === "Paused" || false}
         />
       )}
-    </Box>
+    </div>
   );
 };
 
 export default WDPhaseUI;
+
+// bgcolor: theme.palette[user.member.country]?.light,
