+++ beta-src/src/components/ui/main-screen/RightButton.tsx
@@ -1,0 +1,30 @@
+import React, { FunctionComponent, ReactElement } from "react";
+
+interface RightButtonProps {
+  image: string;
+  text: string;
+  onClick: () => void;
+  className?: string;
+}
+
+const RightButton: FunctionComponent<RightButtonProps> = function ({
+  image,
+  text,
+  onClick,
+  className,
+}): ReactElement {
+  return (
+    <div className={className}>
+      <button onClick={onClick} type="button">
+        <img src={`beta/images/icons/${image}Button.svg`} alt="action" />
+      </button>
+      <div className="bg-black uppercase text-white text-center py-0.5 w-full text-xs font-bold rounded-md">
+        {text}
+      </div>
+    </div>
+  );
+};
+
+RightButton.defaultProps = { className: "" };
+
+export default RightButton;
