+++ beta-src/src/components/ui/main-screen/index.ts
@@ -1,0 +1,5 @@
+/* eslint-disable import/prefer-default-export */
+export { default as TopLeft } from "./TopLeft";
+export { default as TopRight } from "./TopRight";
+export { default as BottomLeft } from "./BottomLeft";
+export { default as BottomRight } from "./BottomRight";
